misc: update missing formats

+++ dex-core/src/main/java/me/darknet/dex/file/items/ClassDataItem.java
@@ -17,10 +17,10 @@ public record ClassDataItem(List<EncodedField> staticFields, List<EncodedField> 
     public static final ItemCodec<ClassDataItem> CODEC = new ItemCodec<>() {
         @Override
         public ClassDataItem read0(Input input, DexMapAccess context) throws IOException {
-            int staticFieldsSize = (int) input.readULeb128();
-            int instanceFieldsSize = (int) input.readULeb128();
-            int directMethodsSize = (int) input.readULeb128();
-            int virtualMethodsSize = (int) input.readULeb128();
+            int staticFieldsSize = input.readULeb128();
+            int instanceFieldsSize = input.readULeb128();
+            int directMethodsSize = input.readULeb128();
+            int virtualMethodsSize = input.readULeb128();
 
             List<EncodedField> staticFields = new ArrayList<>(staticFieldsSize);
             List<EncodedField> instanceFields = new ArrayList<>(instanceFieldsSize);
@@ -29,25 +29,25 @@ public record ClassDataItem(List<EncodedField> staticFields, List<EncodedField> 
 
             int lastFieldIndex = 0;
             for (int i = 0; i < staticFieldsSize; i++) {
-                int fieldIndexDiff = (int) input.readULeb128();
-                int accessFlags = (int) input.readULeb128();
+                int fieldIndexDiff = input.readULeb128();
+                int accessFlags = input.readULeb128();
                 lastFieldIndex += fieldIndexDiff;
                 staticFields.add(new EncodedField(context.fields().get(lastFieldIndex), accessFlags));
             }
 
             lastFieldIndex = 0;
             for (int i = 0; i < instanceFieldsSize; i++) {
-                int fieldIndexDiff = (int) input.readULeb128();
-                int accessFlags = (int) input.readULeb128();
+                int fieldIndexDiff = input.readULeb128();
+                int accessFlags = input.readULeb128();
                 lastFieldIndex += fieldIndexDiff;
                 instanceFields.add(new EncodedField(context.fields().get(lastFieldIndex), accessFlags));
             }
 
             int lastMethodIndex = 0;
             for (int i = 0; i < directMethodsSize; i++) {
-                int methodIndexDiff = (int) input.readULeb128();
-                int accessFlags = (int) input.readULeb128();
-                int codeOffset = (int) input.readULeb128();
+                int methodIndexDiff = input.readULeb128();
+                int accessFlags = input.readULeb128();
+                int codeOffset = input.readULeb128();
                 lastMethodIndex += methodIndexDiff;
                 CodeItem code = codeOffset == 0 ? null : CodeItem.CODEC.read(input.slice(codeOffset), context);
                 directMethods.add(new EncodedMethod(context.methods().get(lastMethodIndex), accessFlags, code));
@@ -55,9 +55,9 @@ public record ClassDataItem(List<EncodedField> staticFields, List<EncodedField> 
 
             lastMethodIndex = 0;
             for (int i = 0; i < virtualMethodsSize; i++) {
-                int methodIndexDiff = (int) input.readULeb128();
-                int accessFlags = (int) input.readULeb128();
-                int codeOffset = (int) input.readULeb128();
+                int methodIndexDiff = input.readULeb128();
+                int accessFlags = input.readULeb128();
+                int codeOffset = input.readULeb128();
                 lastMethodIndex += methodIndexDiff;
                 CodeItem code = codeOffset == 0 ? null : CodeItem.CODEC.read(input.slice(codeOffset), context);
                 virtualMethods.add(new EncodedMethod(context.methods().get(lastMethodIndex), accessFlags, code));
